:sparkles: Permit saving of flight preference orders.
